Added Endless Spawner script
This controls the logic for spawning a particular type of enemy. It needs a list of places to spawn, and a enemy to spawn, in a bundle of three. This is how it is programmed. It may need to be refactored later, but it works for the endless mode we want for sprint 1. The refactoring would just be to make it more reusable, configurable time intervals and the like.