Created NPC script
This script will contain information for the information necessary for the dialog box to exist. It will be part of the UI that is hidden. It will need a prefab above it to show as the character, as well as some data, like a name. This will be provided by the NPC data, which will store a prefab for the character image, a name, and a script.